+++ test/integration/012_deprecation_tests/test_deprecations.py
@@ -24,13 +24,9 @@ class TestDeprecations(BaseTestDeprecations):
         return self.dir("models")
 
     @use_profile('postgres')
-    def test_postgres_deprecations_fail(self):
-        self.run_dbt(strict=True, expect_pass=False)
-
-    @use_profile('postgres')
     def test_postgres_deprecations(self):
         self.assertEqual(deprecations.active_deprecations, set())
-        self.run_dbt(strict=False)
+        self.run_dbt()
         expected = {'adapter:already_exists'}
         self.assertEqual(expected, deprecations.active_deprecations)
 
@@ -48,14 +44,9 @@ class TestMaterializationReturnDeprecation(BaseTestDeprecations):
         }
 
     @use_profile('postgres')
-    def test_postgres_deprecations_fail(self):
-        # this should fail at runtime
-        self.run_dbt(strict=True, expect_pass=False)
-
-    @use_profile('postgres')
     def test_postgres_deprecations(self):
         self.assertEqual(deprecations.active_deprecations, set())
-        self.run_dbt(strict=False)
+        self.run_dbt()
         expected = {'materialization-return'}
         self.assertEqual(expected, deprecations.active_deprecations)
 
@@ -75,23 +66,15 @@ class TestAdapterMacroDeprecation(BaseTestDeprecations):
     @use_profile('postgres')
     def test_postgres_adapter_macro(self):
         self.assertEqual(deprecations.active_deprecations, set())
-        self.run_dbt(strict=False)
+        self.run_dbt()
         expected = {'adapter-macro'}
         self.assertEqual(expected, deprecations.active_deprecations)
-
-    @use_profile('postgres')
-    def test_postgres_adapter_macro_fail(self):
-        self.assertEqual(deprecations.active_deprecations, set())
-        with self.assertRaises(dbt.exceptions.CompilationException) as exc:
-            self.run_dbt(strict=True)
-        exc_str = ' '.join(str(exc.exception).split())  # flatten all whitespace
-        assert 'The "adapter_macro" macro has been deprecated' in exc_str
 
     @use_profile('redshift')
     def test_redshift_adapter_macro(self):
         self.assertEqual(deprecations.active_deprecations, set())
         # pick up the postgres macro
-        self.run_dbt(strict=False)
+        self.run_dbt()
         expected = {'adapter-macro'}
         self.assertEqual(expected, deprecations.active_deprecations)
         
@@ -100,7 +83,7 @@ class TestAdapterMacroDeprecation(BaseTestDeprecations):
         self.assertEqual(deprecations.active_deprecations, set())
         # picked up the default -> error
         with self.assertRaises(dbt.exceptions.CompilationException) as exc:
-            self.run_dbt(strict=False, expect_pass=False)
+            self.run_dbt(expect_pass=False)
         exc_str = ' '.join(str(exc.exception).split())  # flatten all whitespace
         assert 'not allowed' in exc_str  # we saw the default macro
 
@@ -120,24 +103,16 @@ class TestAdapterMacroDeprecationPackages(BaseTestDeprecations):
     @use_profile('postgres')
     def test_postgres_adapter_macro_pkg(self):
         self.assertEqual(deprecations.active_deprecations, set())
-        self.run_dbt(strict=False)
+        self.run_dbt()
         expected = {'adapter-macro'}
         self.assertEqual(expected, deprecations.active_deprecations)
-
-    @use_profile('postgres')
-    def test_postgres_adapter_macro_pkg_fail(self):
-        self.assertEqual(deprecations.active_deprecations, set())
-        with self.assertRaises(dbt.exceptions.CompilationException) as exc:
-            self.run_dbt(strict=True)
-        exc_str = ' '.join(str(exc.exception).split())  # flatten all whitespace
-        assert 'The "adapter_macro" macro has been deprecated' in exc_str
 
     @use_profile('redshift')
     def test_redshift_adapter_macro_pkg(self):
         self.assertEqual(deprecations.active_deprecations, set())
         # pick up the postgres macro
         self.assertEqual(deprecations.active_deprecations, set())
-        self.run_dbt(strict=False)
+        self.run_dbt()
         expected = {'adapter-macro'}
         self.assertEqual(expected, deprecations.active_deprecations)
 
@@ -146,7 +121,7 @@ class TestAdapterMacroDeprecationPackages(BaseTestDeprecations):
         self.assertEqual(deprecations.active_deprecations, set())
         # picked up the default -> error
         with self.assertRaises(dbt.exceptions.CompilationException) as exc:
-            self.run_dbt(strict=False, expect_pass=False)
+            self.run_dbt(expect_pass=False)
         exc_str = ' '.join(str(exc.exception).split())  # flatten all whitespace
         assert 'not allowed' in exc_str  # we saw the default macro
 
@@ -176,14 +151,7 @@ class TestDispatchPackagesDeprecation(BaseTestDeprecations):
     @use_profile('postgres')
     def test_postgres_adapter_macro(self):
         self.assertEqual(deprecations.active_deprecations, set())
-        self.run_dbt(strict=False)
+        self.run_dbt()
         expected = {'dispatch-packages'}
         self.assertEqual(expected, deprecations.active_deprecations)
 
-    @use_profile('postgres')
-    def test_postgres_adapter_macro_fail(self):
-        self.assertEqual(deprecations.active_deprecations, set())
-        with self.assertRaises(dbt.exceptions.CompilationException) as exc:
-            self.run_dbt(strict=True)
-        exc_str = ' '.join(str(exc.exception).split())  # flatten all whitespace
-        assert 'Raised during dispatch for: string_literal' in exc_str